requirements.txt
# Gmail Agent Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# LangGraph / LangChain
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0

# Anthropic Claude
anthropic>=0.18.0

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-cloud-tasks>=2.14.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# CLI
rich>=13.7.0

# Utilities
tenacity>=8.2.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0
freezegun>=1.2.0
testcontainers>=3.7.0
